+++ client/src/components/hub/ContextTable/components/BarSort/BarSort.js
@@ -183,10 +183,8 @@ function BarSort({ sortFields, setSortFields, maxHeight, fields }) {
                   />
                 </div>
                 <div className='BarSort__parameters__box'>
-                  {(
-                    (!!paramFields && Object.keys(paramFields).length > 0)
-                    || (!!metricFields && Object.keys(metricFields).length > 0)
-                  ) ? (
+                  {(!!paramFields && Object.keys(paramFields).length > 0) ||
+                  (!!metricFields && Object.keys(metricFields).length > 0) ? (
                     <>
                       {!!metricFields &&
                         Object.keys(metricFields).length > 0 && (