+++ client/src/utils/index.js
@@ -194,7 +194,7 @@ export function removeOutliers(values, t = 2) {
     if (!values.length) {
       break;
     }
-    if (values[0] === values[values.length-1]) {
+    if (values[0] === values[values.length - 1]) {
       break;
     }
     const q1 = values[Math.floor(values.length / 4)];