+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Changelog
 
+- Add ability to align X axis by relative time and absolute time (roubkar)
+
 ## 1.3.3 Jan 26 2021
 - Improve displaying params of array type in sort box (roubkar)